test: refactor
Separate core tests from vuetify ones.
