+++ .codeclimate.yml
@@ -1,0 +1,4 @@
+version: 2
+exclude_patterns:
+  - node_modules/**
+  - coverage/**
